git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/csl-sol210-3_4-branch@90328 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/testsuite/gcc.dg/i386-sse-5.c
@@ -1,7 +1,7 @@
 /* { dg-do compile { target i?86-*-* } } */
 /* { dg-forbid-option "-m64" } */
 /* { dg-options "-Winline -O2 -march=i386" } */
-typedef int v2df __attribute__ ((mode(V2DF)));
+typedef float v2df __attribute__ ((mode(V2DF)));
 v2df p;
 q(v2df t)
 {			 /* { dg-warning "SSE" "" } */
